+++ src/types/types.ts
@@ -52,6 +52,8 @@ export type Logger = {
   info(message: string): void
   warn(message: string): void
   error(message: string): void
+  debug(message: string): void
+  verbose(message: string): void
 }
 
 export type Context = {